Update in the structure
Enemy movemente changed.
Now exist the Lost condition.
The class Player now have the methos for the movemente instead of being a simple operations in the main funct.
I'm updating the main funct to only inti the other class, so i had to create a class Game to have all the game parameters (still working on it, some bugs were created in the process)