Main Structute Updated
I changed the structure of the main function in order to only initialize the other objects and run the main "while loop" of the game.
I changed the "Game" object to have a method for "GameOver" and "Quitting" the game, those are called in the main function.
I created a "Drawing" object to call all the draw methods from the other objects at the beginning of the main function, and at the end of every main "While Loop"