Collision, parameters adjusted, music
Now exist the collision betwen the player's bullet and the enemy, this will remove the bullet and the enemy from their respective list.

I Changed some parameters of the game as the enemy, player's bullet and player speed and also the way they scale by level.

The music of the game its now on